+++ src/world/entity_group.hpp
@@ -21,7 +21,7 @@ struct entity_group {
    float3 position;
    float rotation_angle = 0.0f;
 
-   int16 layer = 0;
+   int8 layer = 0;
 
    std::vector<object> objects;
    std::vector<light> lights;
