+++ src/world/hintnode.hpp
@@ -50,7 +50,7 @@ enum class hintnode_mode {
 
 struct hintnode {
    std::string name;
-   int16 layer = 0;
+   int8 layer = 0;
    bool hidden = false;
 
    quaternion rotation = {1.0f, 0.0f, 0.0f, 0.0f};
