+++ src/world/object.hpp
@@ -14,7 +14,7 @@ enum class object_class_handle : uint32 {};
 
 struct object {
    std::string name;
-   int16 layer = 0;
+   int8 layer = 0;
    bool hidden = false;
 
    quaternion rotation = {1.0f, 0.0f, 0.0f, 0.0f};
